+++ lib/widgets/dropdown_field.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class DropdownField extends StatelessWidget {
+  final String hint;
+  final String? value;
+  final List<String> items;
+  final Function(String) onChanged;
+
+  const DropdownField({
+    super.key,
+    required this.hint,
+    required this.value,
+    required this.items,
+    required this.onChanged,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: EdgeInsets.symmetric(horizontal: 12),
+      decoration: BoxDecoration(
+        border: Border.all(color: Colors.grey),
+        borderRadius: BorderRadius.circular(4),
+      ),
+      child: DropdownButton<String>(
+        isExpanded: true,
+        underline: SizedBox(),
+        value: value == null || value!.isEmpty ? null : value,
+        hint: Text(hint),
+        items: items.map((String item) {
+          return DropdownMenuItem<String>(value: item, child: Text(item));
+        }).toList(),
+        onChanged: (String? newValue) {
+          if (newValue != null) {
+            onChanged(newValue);
+          }
+        },
+      ),
+    );
+  }
+}
